+++ tests/Serializer/DeserializingConnectionTest.php
@@ -6,12 +6,14 @@ namespace DigitalCraftsman\DeserializingConnection\Serializer;
 
 use DigitalCraftsman\DeserializingConnection\Test\ConnectionTestCase;
 use DigitalCraftsman\DeserializingConnection\Test\DTO\Company;
+use DigitalCraftsman\DeserializingConnection\Test\DTO\Duration;
 use DigitalCraftsman\DeserializingConnection\Test\DTO\User;
 use DigitalCraftsman\DeserializingConnection\Test\ValueObject\CompanyId;
 use DigitalCraftsman\DeserializingConnection\Test\ValueObject\ProjectId;
 use DigitalCraftsman\DeserializingConnection\Test\ValueObject\ProjectIdList;
 use DigitalCraftsman\DeserializingConnection\Test\ValueObject\UserId;
 use DigitalCraftsman\SelfAwareNormalizers\Serializer\ArrayNormalizableNormalizer;
+use DigitalCraftsman\SelfAwareNormalizers\Serializer\IntNormalizableNormalizer;
 use DigitalCraftsman\SelfAwareNormalizers\Serializer\StringNormalizableNormalizer;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
@@ -28,11 +30,14 @@ use Symfony\Component\Serializer\Serializer;
 #[CoversClass(DTO\ResultTransformers::class)]
 #[CoversClass(DTO\ResultTransformer::class)]
 #[CoversClass(Exception\ElementNotFound::class)]
+#[CoversClass(Exception\SingleValueTransformationMustNotContainRenaming::class)]
+#[CoversClass(Exception\IndexMustBeString::class)]
 #[CoversClass(DTO\Exception\ConflictBetweenKeysAndRenameToConfiguration::class)]
 final class DeserializingConnectionTest extends ConnectionTestCase
 {
     private DeserializingConnection $deserializingConnection;
 
+    #[\Override]
     protected function setUp(): void
     {
         parent::setUp();
@@ -42,6 +47,7 @@ final class DeserializingConnectionTest extends ConnectionTestCase
                 normalizers: [
                     new StringNormalizableNormalizer(),
                     new ArrayNormalizableNormalizer(),
+                    new IntNormalizableNormalizer(),
                     new ArrayDenormalizer(),
                     new PropertyNormalizer(
                         propertyTypeExtractor: new PropertyInfoExtractor(
@@ -310,6 +316,200 @@ final class DeserializingConnectionTest extends ConnectionTestCase
     }
 
     #[Test]
+    public function find_one_from_single_value_works(): void
+    {
+        // -- Arrange
+        $userIdString = '417df760-0d16-408f-8201-ec7760dee9fb';
+        $expectedResult = UserId::fromString($userIdString);
+
+        // -- Act
+        $userId = $this->deserializingConnection->findOneFromSingleValue(
+            sql: <<<'SQL'
+                SELECT
+                    '417df760-0d16-408f-8201-ec7760dee9fb'
+                WHERE '417df760-0d16-408f-8201-ec7760dee9fb' = :userId
+                SQL,
+            class: UserId::class,
+            parameters: [
+                'userId' => $userIdString,
+            ],
+        );
+
+        // -- Assert
+        self::assertEquals($expectedResult, $userId);
+    }
+
+    #[Test]
+    public function find_one_from_single_value_works_with_decoding_and_result_transformation(): void
+    {
+        // -- Arrange
+        $expectedResult = new Duration(20);
+
+        // -- Act
+        $duration = $this->deserializingConnection->findOneFromSingleValue(
+            sql: <<<'SQL'
+                SELECT
+                    '15'
+                SQL,
+            class: Duration::class,
+            decoderType: DTO\DecoderType::INT,
+            resultTransformer: DTO\ResultTransformer::toTransform(
+                key: 'key',
+                denormalizeResultToClass: null,
+                transformer: static fn (int $value): int => $value + 5,
+                isTransformedResultNormalized: false,
+            ),
+        );
+
+        // -- Assert
+        self::assertEquals($expectedResult, $duration);
+    }
+
+    #[Test]
+    public function find_one_from_single_value_works_without_results(): void
+    {
+        // -- Act
+        $userId = $this->deserializingConnection->findOneFromSingleValue(
+            sql: <<<'SQL'
+                WITH empty_table AS (
+                    SELECT 1
+                    WHERE false
+                )
+                SELECT *
+                FROM empty_table
+                SQL,
+            class: UserId::class,
+        );
+
+        // -- Assert
+        self::assertNull($userId);
+    }
+
+    #[Test]
+    public function find_one_from_single_value_fails_with_renaming_in_transformation(): void
+    {
+        // -- Assert
+        $this->expectException(Exception\SingleValueTransformationMustNotContainRenaming::class);
+
+        // -- Act
+        $this->deserializingConnection->findOneFromSingleValue(
+            sql: <<<'SQL'
+                WITH empty_table AS (
+                    SELECT 1
+                    WHERE false
+                )
+                SELECT *
+                FROM empty_table
+                SQL,
+            class: UserId::class,
+            resultTransformer: DTO\ResultTransformer::toTransformAndRename(
+                key: 'key',
+                denormalizeResultToClass: null,
+                transformer: static fn (string $name): string => strtoupper($name),
+                isTransformedResultNormalized: false,
+                renameTo: 'renamedKey',
+            ),
+        );
+    }
+
+    #[Test]
+    public function get_one_from_single_value_works(): void
+    {
+        // -- Arrange
+        $userIdString = '417df760-0d16-408f-8201-ec7760dee9fb';
+        $expectedResult = UserId::fromString($userIdString);
+
+        // -- Act
+        $userId = $this->deserializingConnection->getOneFromSingleValue(
+            sql: <<<'SQL'
+                SELECT
+                    '417df760-0d16-408f-8201-ec7760dee9fb'
+                WHERE '417df760-0d16-408f-8201-ec7760dee9fb' = :userId
+                SQL,
+            class: UserId::class,
+            parameters: [
+                'userId' => $userIdString,
+            ],
+        );
+
+        // -- Assert
+        self::assertEquals($expectedResult, $userId);
+    }
+
+    #[Test]
+    public function get_one_from_single_value_works_with_decoding_and_result_transformation(): void
+    {
+        // -- Arrange
+        $expectedResult = new Duration(20);
+
+        // -- Act
+        $duration = $this->deserializingConnection->getOneFromSingleValue(
+            sql: <<<'SQL'
+                SELECT
+                    '15'
+                SQL,
+            class: Duration::class,
+            decoderType: DTO\DecoderType::INT,
+            resultTransformer: DTO\ResultTransformer::toTransform(
+                key: 'key',
+                denormalizeResultToClass: null,
+                transformer: static fn (int $value): int => $value + 5,
+                isTransformedResultNormalized: false,
+            ),
+        );
+
+        // -- Assert
+        self::assertEquals($expectedResult, $duration);
+    }
+
+    #[Test]
+    public function get_one_from_single_value_fails_without_results(): void
+    {
+        // -- Assert
+        $this->expectException(Exception\ElementNotFound::class);
+
+        // -- Act
+        $this->deserializingConnection->getOneFromSingleValue(
+            sql: <<<'SQL'
+                WITH empty_table AS (
+                    SELECT 1
+                    WHERE false
+                )
+                SELECT *
+                FROM empty_table
+                SQL,
+            class: UserId::class,
+        );
+    }
+
+    #[Test]
+    public function get_one_from_single_value_fails_with_renaming_in_transformation(): void
+    {
+        // -- Assert
+        $this->expectException(Exception\SingleValueTransformationMustNotContainRenaming::class);
+
+        // -- Act
+        $this->deserializingConnection->getOneFromSingleValue(
+            sql: <<<'SQL'
+                WITH empty_table AS (
+                    SELECT 1
+                    WHERE false
+                )
+                SELECT *
+                FROM empty_table
+                SQL,
+            class: UserId::class,
+            resultTransformer: DTO\ResultTransformer::toTransformAndRename(
+                key: 'key',
+                denormalizeResultToClass: null,
+                transformer: static fn (string $name): string => strtoupper($name),
+                isTransformedResultNormalized: false,
+                renameTo: 'renamedKey',
+            ),
+        );
+    }
+
+    #[Test]
     public function find_array_works(): void
     {
         // -- Arrange
@@ -362,6 +562,94 @@ final class DeserializingConnectionTest extends ConnectionTestCase
 
         // -- Assert
         self::assertEquals($expectedResult, $users);
+    }
+
+    #[Test]
+    public function find_array_works_with_indexed_by(): void
+    {
+        // -- Arrange
+        $expectedResult = [
+            '417df760-0d16-408f-8201-ec7760dee9fb' => new User(
+                userId: UserId::fromString('417df760-0d16-408f-8201-ec7760dee9fb'),
+                name: 'JOHN DOE',
+                accessibleProjects: new ProjectIdList([
+                    ProjectId::fromString('05f620c2-ea64-4012-816f-884310f69dd0'),
+                    ProjectId::fromString('91f47435-208d-4344-990b-ae17bd4b13fa'),
+                ]),
+                companies: [],
+            ),
+            'ef64a500-db7b-49a8-b670-8eca24936688' => new User(
+                userId: UserId::fromString('ef64a500-db7b-49a8-b670-8eca24936688'),
+                name: 'JANE DOE',
+                accessibleProjects: ProjectIdList::emptyList(),
+                companies: [],
+            ),
+        ];
+
+        // -- Act
+        $users = $this->deserializingConnection->findArray(
+            sql: <<<'SQL'
+                SELECT
+                        user_id AS "userId",
+                        name,
+                        accessible_projects AS "accessibleProjects",
+                        '[]' AS "companies"
+                    FROM (
+                        VALUES
+                            ('417df760-0d16-408f-8201-ec7760dee9fb', 'John Doe', '["05f620c2-ea64-4012-816f-884310f69dd0", "91f47435-208d-4344-990b-ae17bd4b13fa"]'),
+                            ('ef64a500-db7b-49a8-b670-8eca24936688', 'Jane Doe', '[]')
+                    ) AS users(user_id, name, accessible_projects)
+                SQL,
+            class: User::class,
+            decoderTypes: [
+                'accessibleProjects' => DTO\DecoderType::JSON,
+                'companies' => DTO\DecoderType::JSON,
+            ],
+            resultTransformers: [
+                DTO\ResultTransformer::toTransform(
+                    key: 'name',
+                    denormalizeResultToClass: null,
+                    transformer: static fn (string $name): string => strtoupper($name),
+                    isTransformedResultNormalized: false,
+                ),
+            ],
+            indexedBy: static fn (User $user): string => (string) $user->userId,
+        );
+
+        // -- Assert
+        self::assertEquals($expectedResult, $users);
+    }
+
+    #[Test]
+    public function find_array_fails_with_invalid_index_generation(): void
+    {
+        // -- Assert
+        $this->expectException(Exception\IndexMustBeString::class);
+
+        // -- Act
+        /**
+         * @psalm-suppress InvalidArgument Provided on purpose to test the exception
+         */
+        $this->deserializingConnection->findArray(
+            sql: <<<'SQL'
+                SELECT
+                        user_id AS "userId",
+                        name,
+                        accessible_projects AS "accessibleProjects",
+                        '[]' AS "companies"
+                    FROM (
+                        VALUES
+                            ('417df760-0d16-408f-8201-ec7760dee9fb', 'John Doe', '["05f620c2-ea64-4012-816f-884310f69dd0", "91f47435-208d-4344-990b-ae17bd4b13fa"]'),
+                            ('ef64a500-db7b-49a8-b670-8eca24936688', 'Jane Doe', '[]')
+                    ) AS users(user_id, name, accessible_projects)
+                SQL,
+            class: User::class,
+            decoderTypes: [
+                'accessibleProjects' => DTO\DecoderType::JSON,
+                'companies' => DTO\DecoderType::JSON,
+            ],
+            indexedBy: static fn (User $user): int => 15,
+        );
     }
 
     #[Test]